+++ legacy_packages/react/src/components/QRCode.tsx
@@ -102,7 +102,7 @@ export const PlaceholderKeyframes = keyframes`
   100%{ background-position: -100% 0; }
 `;
 
-export const QRPlaceholder = /* @__PURE__ */ StyledDiv(() => {
+export const QRPlaceholder = /* @__PURE__ */ StyledDiv((_) => {
   const theme = useCustomTheme();
   return {
     "--color": theme.colors.skeletonBg,