+++ legacy_packages/react/src/wallet/ConnectWallet/SignatureScreen.tsx
@@ -368,7 +368,7 @@ const plusAnimation = keyframes`
 }
 `;
 
-const PulsatingContainer = /* @__PURE__ */ StyledDiv(() => {
+const PulsatingContainer = /* @__PURE__ */ StyledDiv((_) => {
   const theme = useCustomTheme();
   return {
     position: "relative",