+++ legacy_packages/storage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @thirdweb-dev/storage
 
+## 2.0.16
+
+### Patch Changes
+
+- [#3414](https://github.com/thirdweb-dev/js/pull/3414) [`5ca9094`](https://github.com/thirdweb-dev/js/commit/5ca9094e260e9758f2e8964617e172ae542306fc) Thanks [@jnsdls](https://github.com/jnsdls)! - update dependencies
+
+- Updated dependencies [[`5ca9094`](https://github.com/thirdweb-dev/js/commit/5ca9094e260e9758f2e8964617e172ae542306fc)]:
+  - @thirdweb-dev/crypto@0.2.7
+
 ## 2.0.15
 
 ### Patch Changes